+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## Fixed
 
+- Use tools.namespace.track for tracking/reloading namespaces in watch mode.
+  This should make this a lot more reliable.
+- Change the `add-classpath` classloader hack so it doesn't mess up the thread
+  binding stack.
+- Make dots reporter compatible with newer versions of matcher-combinators.
+
 ## Changed
 
 # 0.0-138 (2018-09-17 / 9bc74e4)